Double precision for gameclock.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2990 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/client.h
@@ -584,7 +584,7 @@ typedef struct
 	qboolean sendprespawn;
 	int contentstage;
 
-	float ktprogametime;
+	double ktprogametime;
 	enum {
 		KTPRO_DONTKNOW,
 		KTPRO_COUNTDOWN,
